Create facebook_post_scraper_simple.py
This script will download all Facebook posts, and its metadata, and save it as a .csv and .json file. It will NOT download any photos or videos associated with the post.

It is recommended NOT to use cookies to download FB posts. Only use cookies if Facebook is temporarily banning/rate limiting you.

Features:

1. An "archive" csv file is generated, for easy tracking of which accounts you have backed up, when you last backed it up, etc
2. Scraper will not pull posts that have already been pulled before, to avoid FB rate limiting (to disable this one-time, simply delete the row in the .csv file for the page you are backing up.)
3. Can pull posts using a proxy
4. Can use FB account cookies.